Detection of outliers using quartiles and by comparing with other data in same or previous period.

Detection of outliers using quartiles and by comparing with other data in same or previous period.

Usage

Quartile(
  data,
  id,
  x1,
  y1,
  x2 = NULL,
  y2 = NULL,
  strataName = NULL,
  pKL = 1.5,
  pKU = 1.5
)

Arguments

data

Input of Quartile is a data set of class data.frame.

id

Name of the identification variable.

x1

Name of x variable in period t.

y1

Name of y variable in period t.

x2

Name of x variable in period t-1. Optional

y2

Name of y variable in period t-1. Optional

strataName

Name of the stratification variable. Optional

pKL

Parameter for lower limit.

pKU

Parameter for upper limit.

Value

Output of Quartile is a data set of class data.frame. Only units with both x1 and y1 not missing and greater than zero are included. The variables are:

id

The input identification variable

x1

The input x1 variable

y1

The input y1 variable

x2

The input x2 variable

y2

The input y2 variable

ratio

The ratio between x1 and y1

ratio2

The ratio between x2 and y2

ratioAll

The ratio between the sum of x1 and the sum of y1 aggregated over the whole data set

ratioAll2

The ratio between the sum of x2 and the sum of y2 aggregated over the whole data set

ratioStr

The ratio between the sum of x1 and the sum of y1 aggregated over the stratum

ratioStr2

The ratio between the sum of x2 and the sum of y2 aggregated over the stratum

lowerLimit

The lower limit of the ratio

upperLimit

The upper limit of the ratio

outlier

A binary variable indicating whether the observation is outside the limits [q1 - pKL*(M - q1), q3 + pKU*(q3 - M)]

, where M is the median and q1 and q3 the 1st an 3rd quartile respectively.

strata

Strata name or number

ranking

The rank of ratio. For plotting purposes

References

Datarevisjon: Kontroll, granskning og retting av data. Anbefalt praksis, Statistisk sentralbyrås håndbøker, 2005/84.

Author

Magnar Lillegård

Examples

 minedata <- KostraData("testdata")
 minedata$strata <- as.character(c(rep(1, 61), rep(2, 91), rep(3, 98), rep(4, 81), rep(5, 85)))
 Quartile(data = minedata, id = "Region", x1 = "areal_130_eier_2015", y1 = "areal_130_leier_2015", pKL = 2, pKU = 2)
#>        id    x1    y1 x2 y2      ratio ratio2 ratioAll ratioAll2 ratioStr
#> 1   10100  3956  8365 NA NA  0.4729229     NA 2.534378        NA       NA
#> 2   10600 25306  4040 NA NA  6.2638614     NA 2.534378        NA       NA
#> 3   12200  4867   200 NA NA 24.3350000     NA 2.534378        NA       NA
#> 4   12800  5252   632 NA NA  8.3101266     NA 2.534378        NA       NA
#> 5   13600  5648  1779 NA NA  3.1748173     NA 2.534378        NA       NA
#> 6   21400  2516   891 NA NA  2.8237935     NA 2.534378        NA       NA
#> 7   21500  4877  1277 NA NA  3.8191073     NA 2.534378        NA       NA
#> 8   21700 14450   937 NA NA 15.4215582     NA 2.534378        NA       NA
#> 9   21900 28015  2814 NA NA  9.9555792     NA 2.534378        NA       NA
#> 10  22000 16708  2762 NA NA  6.0492397     NA 2.534378        NA       NA
#> 11  23000 13287   246 NA NA 54.0121951     NA 2.534378        NA       NA
#> 12  23600  8683   339 NA NA 25.6135693     NA 2.534378        NA       NA
#> 13  30100  9881 20904 NA NA  0.4726847     NA 2.534378        NA       NA
#> 14  40200  7211  3634 NA NA  1.9843148     NA 2.534378        NA       NA
#> 15  41700  7645  1855 NA NA  4.1212938     NA 2.534378        NA       NA
#> 16  42600  3780    40 NA NA 94.5000000     NA 2.534378        NA       NA
#> 17  42700 15973  3459 NA NA  4.6178086     NA 2.534378        NA       NA
#> 18  50100  8724  1279 NA NA  6.8209539     NA 2.534378        NA       NA
#> 19  52900  6021  2750 NA NA  2.1894545     NA 2.534378        NA       NA
#> 20  60500  9126  5336 NA NA  1.7102699     NA 2.534378        NA       NA
#> 21  61900  3174   445 NA NA  7.1325843     NA 2.534378        NA       NA
#> 22  62400 19400   350 NA NA 55.4285714     NA 2.534378        NA       NA
#> 23  62600  6908  5061 NA NA  1.3649476     NA 2.534378        NA       NA
#> 24  63300  2548   170 NA NA 14.9882353     NA 2.534378        NA       NA
#> 25  70200  4302  1392 NA NA  3.0905172     NA 2.534378        NA       NA
#> 26  70600 10255   770 NA NA 13.3181818     NA 2.534378        NA       NA
#> 27  70900  2234  2807 NA NA  0.7958675     NA 2.534378        NA       NA
#> 28  72000   150  4466 NA NA  0.0335871     NA 2.534378        NA       NA
#> 29  80500 16941   300 NA NA 56.4700000     NA 2.534378        NA       NA
#> 30  80600  8977 20495 NA NA  0.4380093     NA 2.534378        NA       NA
#> 31  81500  1888  2040 NA NA  0.9254902     NA 2.534378        NA       NA
#> 32  82100  2371  1411 NA NA  1.6803685     NA 2.534378        NA       NA
#> 33  82600  5401   815 NA NA  6.6269939     NA 2.534378        NA       NA
#> 34  83000  1670   250 NA NA  6.6800000     NA 2.534378        NA       NA
#> 35  83400  2600   400 NA NA  6.5000000     NA 2.534378        NA       NA
#> 36  91900  2986   150 NA NA 19.9066667     NA 2.534378        NA       NA
#> 37  93800   540  2224 NA NA  0.2428058     NA 2.534378        NA       NA
#> 38 100100 21802 13728 NA NA  1.5881410     NA 2.534378        NA       NA
#> 39 110100  2744  2670 NA NA  1.0277154     NA 2.534378        NA       NA
#> 40 110200 33029  9991 NA NA  3.3058753     NA 2.534378        NA       NA
#> 41 110300 18953  5185 NA NA  3.6553520     NA 2.534378        NA       NA
#> 42 110600  8499  8938 NA NA  0.9508839     NA 2.534378        NA       NA
#> 43 112400 12100  1200 NA NA 10.0833333     NA 2.534378        NA       NA
#> 44 112900  1647    18 NA NA 91.5000000     NA 2.534378        NA       NA
#> 45 113000  5947  2815 NA NA  2.1126110     NA 2.534378        NA       NA
#> 46 116000  5023   450 NA NA 11.1622222     NA 2.534378        NA       NA
#> 47 120100 50336 24755 NA NA  2.0333670     NA 2.534378        NA       NA
#> 48 122400  6134  2245 NA NA  2.7322940     NA 2.534378        NA       NA
#> 49 122800  2480   929 NA NA  2.6695371     NA 2.534378        NA       NA
#> 50 123500  7161  1357 NA NA  5.2770818     NA 2.534378        NA       NA
#> 51 124500  1748  1238 NA NA  1.4119548     NA 2.534378        NA       NA
#> 52 124700  9399   244 NA NA 38.5204918     NA 2.534378        NA       NA
#> 53 126300  1948  2654 NA NA  0.7339864     NA 2.534378        NA       NA
#> 54 126600  1800   300 NA NA  6.0000000     NA 2.534378        NA       NA
#> 55 140100  7279  3004 NA NA  2.4231025     NA 2.534378        NA       NA
#> 56 141600  3384   474 NA NA  7.1392405     NA 2.534378        NA       NA
#> 57 143900  4057  2132 NA NA  1.9029081     NA 2.534378        NA       NA
#> 58 150500  5040  5999 NA NA  0.8401400     NA 2.534378        NA       NA
#> 59 151400  1607   367 NA NA  4.3787466     NA 2.534378        NA       NA
#> 60 152800  2146   210 NA NA 10.2190476     NA 2.534378        NA       NA
#> 61 153400  3302   506 NA NA  6.5256917     NA 2.534378        NA       NA
#> 62 157600  2130   350 NA NA  6.0857143     NA 2.534378        NA       NA
#> 63 160100 14022 13517 NA NA  1.0373604     NA 2.534378        NA       NA
#> 64 161200   965   656 NA NA  1.4710366     NA 2.534378        NA       NA
#> 65 161700  4611   180 NA NA 25.6166667     NA 2.534378        NA       NA
#> 66 162000  3621  3000 NA NA  1.2070000     NA 2.534378        NA       NA
#> 67 162700  2194   176 NA NA 12.4659091     NA 2.534378        NA       NA
#> 68 164000  1794   165 NA NA 10.8727273     NA 2.534378        NA       NA
#> 69 164400   660   350 NA NA  1.8857143     NA 2.534378        NA       NA
#> 70 166300  3427   583 NA NA  5.8782161     NA 2.534378        NA       NA
#> 71 170300  6563  2652 NA NA  2.4747360     NA 2.534378        NA       NA
#> 72 171400  7495  1926 NA NA  3.8914849     NA 2.534378        NA       NA
#> 73 173600  1995   100 NA NA 19.9500000     NA 2.534378        NA       NA
#> 74 175000   910  2000 NA NA  0.4550000     NA 2.534378        NA       NA
#> 75 180400 15448  5273 NA NA  2.9296416     NA 2.534378        NA       NA
#> 76 182400  8200   980 NA NA  8.3673469     NA 2.534378        NA       NA
#> 77 183300  4833   484 NA NA  9.9855372     NA 2.534378        NA       NA
#> 78 183600  1173   370 NA NA  3.1702703     NA 2.534378        NA       NA
#> 79 184000  1904  3121 NA NA  0.6100609     NA 2.534378        NA       NA
#> 80 184100  2697  1826 NA NA  1.4769989     NA 2.534378        NA       NA
#> 81 186800  3990   150 NA NA 26.6000000     NA 2.534378        NA       NA
#> 82 193800  1935  1695 NA NA  1.1415929     NA 2.534378        NA       NA
#> 83 202400   964    25 NA NA 38.5600000     NA 2.534378        NA       NA
#>    ratioStr2 lowerLimit upperLimit outlier strata ranking
#> 1         NA  -2.880206   22.32034       0      1       6
#> 2         NA  -2.880206   22.32034       0      1      51
#> 3         NA  -2.880206   22.32034       1      1      73
#> 4         NA  -2.880206   22.32034       0      1      59
#> 5         NA  -2.880206   22.32034       0      1      38
#> 6         NA  -2.880206   22.32034       0      1      34
#> 7         NA  -2.880206   22.32034       0      1      41
#> 8         NA  -2.880206   22.32034       0      1      70
#> 9         NA  -2.880206   22.32034       0      1      61
#> 10        NA  -2.880206   22.32034       0      1      49
#> 11        NA  -2.880206   22.32034       1      1      79
#> 12        NA  -2.880206   22.32034       1      1      74
#> 13        NA  -2.880206   22.32034       0      1       5
#> 14        NA  -2.880206   22.32034       0      1      26
#> 15        NA  -2.880206   22.32034       0      1      43
#> 16        NA  -2.880206   22.32034       1      1      83
#> 17        NA  -2.880206   22.32034       0      1      45
#> 18        NA  -2.880206   22.32034       0      1      56
#> 19        NA  -2.880206   22.32034       0      1      29
#> 20        NA  -2.880206   22.32034       0      1      23
#> 21        NA  -2.880206   22.32034       0      1      57
#> 22        NA  -2.880206   22.32034       1      1      80
#> 23        NA  -2.880206   22.32034       0      1      17
#> 24        NA  -2.880206   22.32034       0      1      69
#> 25        NA  -2.880206   22.32034       0      1      36
#> 26        NA  -2.880206   22.32034       0      1      68
#> 27        NA  -2.880206   22.32034       0      1       9
#> 28        NA  -2.880206   22.32034       0      1       1
#> 29        NA  -2.880206   22.32034       1      1      81
#> 30        NA  -2.880206   22.32034       0      1       3
#> 31        NA  -2.880206   22.32034       0      1      11
#> 32        NA  -2.880206   22.32034       0      1      22
#> 33        NA  -2.880206   22.32034       0      1      54
#> 34        NA  -2.880206   22.32034       0      1      55
#> 35        NA  -2.880206   22.32034       0      1      52
#> 36        NA  -2.880206   22.32034       0      1      71
#> 37        NA  -2.880206   22.32034       0      1       2
#> 38        NA  -2.880206   22.32034       0      1      21
#> 39        NA  -2.880206   22.32034       0      1      13
#> 40        NA  -2.880206   22.32034       0      1      39
#> 41        NA  -2.880206   22.32034       0      1      40
#> 42        NA  -2.880206   22.32034       0      1      12
#> 43        NA  -2.880206   22.32034       0      1      63
#> 44        NA  -2.880206   22.32034       1      1      82
#> 45        NA  -2.880206   22.32034       0      1      28
#> 46        NA  -2.880206   22.32034       0      1      66
#> 47        NA  -2.880206   22.32034       0      1      27
#> 48        NA  -2.880206   22.32034       0      1      33
#> 49        NA  -2.880206   22.32034       0      1      32
#> 50        NA  -2.880206   22.32034       0      1      46
#> 51        NA  -2.880206   22.32034       0      1      18
#> 52        NA  -2.880206   22.32034       1      1      77
#> 53        NA  -2.880206   22.32034       0      1       8
#> 54        NA  -2.880206   22.32034       0      1      48
#> 55        NA  -2.880206   22.32034       0      1      30
#> 56        NA  -2.880206   22.32034       0      1      58
#> 57        NA  -2.880206   22.32034       0      1      25
#> 58        NA  -2.880206   22.32034       0      1      10
#> 59        NA  -2.880206   22.32034       0      1      44
#> 60        NA  -2.880206   22.32034       0      1      64
#> 61        NA  -2.880206   22.32034       0      1      53
#> 62        NA  -2.880206   22.32034       0      1      50
#> 63        NA  -2.880206   22.32034       0      1      14
#> 64        NA  -2.880206   22.32034       0      1      19
#> 65        NA  -2.880206   22.32034       1      1      75
#> 66        NA  -2.880206   22.32034       0      1      16
#> 67        NA  -2.880206   22.32034       0      1      67
#> 68        NA  -2.880206   22.32034       0      1      65
#> 69        NA  -2.880206   22.32034       0      1      24
#> 70        NA  -2.880206   22.32034       0      1      47
#> 71        NA  -2.880206   22.32034       0      1      31
#> 72        NA  -2.880206   22.32034       0      1      42
#> 73        NA  -2.880206   22.32034       0      1      72
#> 74        NA  -2.880206   22.32034       0      1       4
#> 75        NA  -2.880206   22.32034       0      1      35
#> 76        NA  -2.880206   22.32034       0      1      60
#> 77        NA  -2.880206   22.32034       0      1      62
#> 78        NA  -2.880206   22.32034       0      1      37
#> 79        NA  -2.880206   22.32034       0      1       7
#> 80        NA  -2.880206   22.32034       0      1      20
#> 81        NA  -2.880206   22.32034       1      1      76
#> 82        NA  -2.880206   22.32034       0      1      15
#> 83        NA  -2.880206   22.32034       1      1      78
 Quartile(data = minedata, id = "Region", x1 = "areal_130_eier_2015", y1 = "areal_130_leier_2015", strataName = "strata")
#>        id    x1    y1 x2 y2      ratio ratio2 ratioAll ratioAll2 ratioStr
#> 1   10100  3956  8365 NA NA  0.4729229     NA 2.534378        NA 3.286724
#> 2   10600 25306  4040 NA NA  6.2638614     NA 2.534378        NA 3.286724
#> 3   12200  4867   200 NA NA 24.3350000     NA 2.534378        NA 3.286724
#> 4   12800  5252   632 NA NA  8.3101266     NA 2.534378        NA 3.286724
#> 5   13600  5648  1779 NA NA  3.1748173     NA 2.534378        NA 3.286724
#> 6   21400  2516   891 NA NA  2.8237935     NA 2.534378        NA 3.286724
#> 7   21500  4877  1277 NA NA  3.8191073     NA 2.534378        NA 3.286724
#> 8   21700 14450   937 NA NA 15.4215582     NA 2.534378        NA 3.286724
#> 9   21900 28015  2814 NA NA  9.9555792     NA 2.534378        NA 3.286724
#> 10  22000 16708  2762 NA NA  6.0492397     NA 2.534378        NA 3.286724
#> 11  23000 13287   246 NA NA 54.0121951     NA 2.534378        NA 3.286724
#> 12  23600  8683   339 NA NA 25.6135693     NA 2.534378        NA 3.286724
#> 13  30100  9881 20904 NA NA  0.4726847     NA 2.534378        NA 3.286724
#> 14  40200  7211  3634 NA NA  1.9843148     NA 2.534378        NA 3.286724
#> 15  41700  7645  1855 NA NA  4.1212938     NA 2.534378        NA 3.286724
#> 16  42600  3780    40 NA NA 94.5000000     NA 2.534378        NA 3.286724
#> 17  42700 15973  3459 NA NA  4.6178086     NA 2.534378        NA 3.286724
#> 18  50100  8724  1279 NA NA  6.8209539     NA 2.534378        NA 2.196443
#> 19  52900  6021  2750 NA NA  2.1894545     NA 2.534378        NA 2.196443
#> 20  60500  9126  5336 NA NA  1.7102699     NA 2.534378        NA 2.196443
#> 21  61900  3174   445 NA NA  7.1325843     NA 2.534378        NA 2.196443
#> 22  62400 19400   350 NA NA 55.4285714     NA 2.534378        NA 2.196443
#> 23  62600  6908  5061 NA NA  1.3649476     NA 2.534378        NA 2.196443
#> 24  63300  2548   170 NA NA 14.9882353     NA 2.534378        NA 2.196443
#> 25  70200  4302  1392 NA NA  3.0905172     NA 2.534378        NA 2.196443
#> 26  70600 10255   770 NA NA 13.3181818     NA 2.534378        NA 2.196443
#> 27  70900  2234  2807 NA NA  0.7958675     NA 2.534378        NA 2.196443
#> 28  72000   150  4466 NA NA  0.0335871     NA 2.534378        NA 2.196443
#> 29  80500 16941   300 NA NA 56.4700000     NA 2.534378        NA 2.196443
#> 30  80600  8977 20495 NA NA  0.4380093     NA 2.534378        NA 2.196443
#> 31  81500  1888  2040 NA NA  0.9254902     NA 2.534378        NA 2.196443
#> 32  82100  2371  1411 NA NA  1.6803685     NA 2.534378        NA 2.196443
#> 33  82600  5401   815 NA NA  6.6269939     NA 2.534378        NA 2.196443
#> 34  83000  1670   250 NA NA  6.6800000     NA 2.534378        NA 2.196443
#> 35  83400  2600   400 NA NA  6.5000000     NA 2.534378        NA 2.196443
#> 36  91900  2986   150 NA NA 19.9066667     NA 2.534378        NA 2.196443
#> 37  93800   540  2224 NA NA  0.2428058     NA 2.534378        NA 2.196443
#> 38 100100 21802 13728 NA NA  1.5881410     NA 2.534378        NA 2.436586
#> 39 110100  2744  2670 NA NA  1.0277154     NA 2.534378        NA 2.436586
#> 40 110200 33029  9991 NA NA  3.3058753     NA 2.534378        NA 2.436586
#> 41 110300 18953  5185 NA NA  3.6553520     NA 2.534378        NA 2.436586
#> 42 110600  8499  8938 NA NA  0.9508839     NA 2.534378        NA 2.436586
#> 43 112400 12100  1200 NA NA 10.0833333     NA 2.534378        NA 2.436586
#> 44 112900  1647    18 NA NA 91.5000000     NA 2.534378        NA 2.436586
#> 45 113000  5947  2815 NA NA  2.1126110     NA 2.534378        NA 2.436586
#> 46 116000  5023   450 NA NA 11.1622222     NA 2.534378        NA 2.436586
#> 47 120100 50336 24755 NA NA  2.0333670     NA 2.534378        NA 2.436586
#> 48 122400  6134  2245 NA NA  2.7322940     NA 2.534378        NA 2.436586
#> 49 122800  2480   929 NA NA  2.6695371     NA 2.534378        NA 2.436586
#> 50 123500  7161  1357 NA NA  5.2770818     NA 2.534378        NA 2.436586
#> 51 124500  1748  1238 NA NA  1.4119548     NA 2.534378        NA 2.436586
#> 52 124700  9399   244 NA NA 38.5204918     NA 2.534378        NA 2.436586
#> 53 126300  1948  2654 NA NA  0.7339864     NA 2.534378        NA 2.436586
#> 54 126600  1800   300 NA NA  6.0000000     NA 2.534378        NA 2.436586
#> 55 140100  7279  3004 NA NA  2.4231025     NA 2.534378        NA 2.436586
#> 56 141600  3384   474 NA NA  7.1392405     NA 2.534378        NA 2.436586
#> 57 143900  4057  2132 NA NA  1.9029081     NA 2.534378        NA 2.436586
#> 58 150500  5040  5999 NA NA  0.8401400     NA 2.534378        NA 1.908605
#> 59 151400  1607   367 NA NA  4.3787466     NA 2.534378        NA 1.908605
#> 60 152800  2146   210 NA NA 10.2190476     NA 2.534378        NA 1.908605
#> 61 153400  3302   506 NA NA  6.5256917     NA 2.534378        NA 1.908605
#> 62 157600  2130   350 NA NA  6.0857143     NA 2.534378        NA 1.908605
#> 63 160100 14022 13517 NA NA  1.0373604     NA 2.534378        NA 1.908605
#> 64 161200   965   656 NA NA  1.4710366     NA 2.534378        NA 1.908605
#> 65 161700  4611   180 NA NA 25.6166667     NA 2.534378        NA 1.908605
#> 66 162000  3621  3000 NA NA  1.2070000     NA 2.534378        NA 1.908605
#> 67 162700  2194   176 NA NA 12.4659091     NA 2.534378        NA 1.908605
#> 68 164000  1794   165 NA NA 10.8727273     NA 2.534378        NA 1.908605
#> 69 164400   660   350 NA NA  1.8857143     NA 2.534378        NA 1.908605
#> 70 166300  3427   583 NA NA  5.8782161     NA 2.534378        NA 1.908605
#> 71 170300  6563  2652 NA NA  2.4747360     NA 2.534378        NA 1.908605
#> 72 171400  7495  1926 NA NA  3.8914849     NA 2.534378        NA 1.908605
#> 73 173600  1995   100 NA NA 19.9500000     NA 2.534378        NA 1.908605
#> 74 175000   910  2000 NA NA  0.4550000     NA 2.534378        NA 1.908605
#> 75 180400 15448  5273 NA NA  2.9296416     NA 2.534378        NA 2.954898
#> 76 182400  8200   980 NA NA  8.3673469     NA 2.534378        NA 2.954898
#> 77 183300  4833   484 NA NA  9.9855372     NA 2.534378        NA 2.954898
#> 78 183600  1173   370 NA NA  3.1702703     NA 2.534378        NA 2.954898
#> 79 184000  1904  3121 NA NA  0.6100609     NA 2.534378        NA 2.954898
#> 80 184100  2697  1826 NA NA  1.4769989     NA 2.534378        NA 2.954898
#> 81 186800  3990   150 NA NA 26.6000000     NA 2.534378        NA 2.954898
#> 82 193800  1935  1695 NA NA  1.1415929     NA 2.534378        NA 2.954898
#> 83 202400   964    25 NA NA 38.5600000     NA 2.534378        NA 2.954898
#>    ratioStr2 lowerLimit upperLimit outlier strata ranking
#> 1         NA   -1.75159   19.24886       0      1       6
#> 2         NA   -1.75159   19.24886       0      1      51
#> 3         NA   -1.75159   19.24886       1      1      73
#> 4         NA   -1.75159   19.24886       0      1      59
#> 5         NA   -1.75159   19.24886       0      1      38
#> 6         NA   -1.75159   19.24886       0      1      34
#> 7         NA   -1.75159   19.24886       0      1      41
#> 8         NA   -1.75159   19.24886       0      1      70
#> 9         NA   -1.75159   19.24886       0      1      61
#> 10        NA   -1.75159   19.24886       0      1      49
#> 11        NA   -1.75159   19.24886       1      1      79
#> 12        NA   -1.75159   19.24886       1      1      74
#> 13        NA   -1.75159   19.24886       0      1       5
#> 14        NA   -1.75159   19.24886       0      1      26
#> 15        NA   -1.75159   19.24886       0      1      43
#> 16        NA   -1.75159   19.24886       1      1      83
#> 17        NA   -1.75159   19.24886       0      1      45
#> 18        NA   -1.75159   19.24886       0      2      56
#> 19        NA   -1.75159   19.24886       0      2      29
#> 20        NA   -1.75159   19.24886       0      2      23
#> 21        NA   -1.75159   19.24886       0      2      57
#> 22        NA   -1.75159   19.24886       1      2      80
#> 23        NA   -1.75159   19.24886       0      2      17
#> 24        NA   -1.75159   19.24886       0      2      69
#> 25        NA   -1.75159   19.24886       0      2      36
#> 26        NA   -1.75159   19.24886       0      2      68
#> 27        NA   -1.75159   19.24886       0      2       9
#> 28        NA   -1.75159   19.24886       0      2       1
#> 29        NA   -1.75159   19.24886       1      2      81
#> 30        NA   -1.75159   19.24886       0      2       3
#> 31        NA   -1.75159   19.24886       0      2      11
#> 32        NA   -1.75159   19.24886       0      2      22
#> 33        NA   -1.75159   19.24886       0      2      54
#> 34        NA   -1.75159   19.24886       0      2      55
#> 35        NA   -1.75159   19.24886       0      2      52
#> 36        NA   -1.75159   19.24886       1      2      71
#> 37        NA   -1.75159   19.24886       0      2       2
#> 38        NA   -1.75159   19.24886       0      3      21
#> 39        NA   -1.75159   19.24886       0      3      13
#> 40        NA   -1.75159   19.24886       0      3      39
#> 41        NA   -1.75159   19.24886       0      3      40
#> 42        NA   -1.75159   19.24886       0      3      12
#> 43        NA   -1.75159   19.24886       0      3      63
#> 44        NA   -1.75159   19.24886       1      3      82
#> 45        NA   -1.75159   19.24886       0      3      28
#> 46        NA   -1.75159   19.24886       0      3      66
#> 47        NA   -1.75159   19.24886       0      3      27
#> 48        NA   -1.75159   19.24886       0      3      33
#> 49        NA   -1.75159   19.24886       0      3      32
#> 50        NA   -1.75159   19.24886       0      3      46
#> 51        NA   -1.75159   19.24886       0      3      18
#> 52        NA   -1.75159   19.24886       1      3      77
#> 53        NA   -1.75159   19.24886       0      3       8
#> 54        NA   -1.75159   19.24886       0      3      48
#> 55        NA   -1.75159   19.24886       0      3      30
#> 56        NA   -1.75159   19.24886       0      3      58
#> 57        NA   -1.75159   19.24886       0      3      25
#> 58        NA   -1.75159   19.24886       0      4      10
#> 59        NA   -1.75159   19.24886       0      4      44
#> 60        NA   -1.75159   19.24886       0      4      64
#> 61        NA   -1.75159   19.24886       0      4      53
#> 62        NA   -1.75159   19.24886       0      4      50
#> 63        NA   -1.75159   19.24886       0      4      14
#> 64        NA   -1.75159   19.24886       0      4      19
#> 65        NA   -1.75159   19.24886       1      4      75
#> 66        NA   -1.75159   19.24886       0      4      16
#> 67        NA   -1.75159   19.24886       0      4      67
#> 68        NA   -1.75159   19.24886       0      4      65
#> 69        NA   -1.75159   19.24886       0      4      24
#> 70        NA   -1.75159   19.24886       0      4      47
#> 71        NA   -1.75159   19.24886       0      4      31
#> 72        NA   -1.75159   19.24886       0      4      42
#> 73        NA   -1.75159   19.24886       1      4      72
#> 74        NA   -1.75159   19.24886       0      4       4
#> 75        NA   -1.75159   19.24886       0      5      35
#> 76        NA   -1.75159   19.24886       0      5      60
#> 77        NA   -1.75159   19.24886       0      5      62
#> 78        NA   -1.75159   19.24886       0      5      37
#> 79        NA   -1.75159   19.24886       0      5       7
#> 80        NA   -1.75159   19.24886       0      5      20
#> 81        NA   -1.75159   19.24886       1      5      76
#> 82        NA   -1.75159   19.24886       0      5      15
#> 83        NA   -1.75159   19.24886       1      5      78
 Quartile(data = minedata, id = "Region", x1 = "areal_130_eier_2015", y1 = "areal_130_leier_2015",
          x2 = "areal_130_eier_2014", y2 = "areal_130_leier_2014", strataName = "strata")
#>        id    x1    y1    x2    y2      ratio     ratio2 ratioAll ratioAll2
#> 1   10100  3956  8365     0 15738  0.4729229  0.0000000 2.534378   2.52813
#> 2   10600 25306  4040 25306  2863  6.2638614  8.8389801 2.534378   2.52813
#> 3   12200  4867   200  4867    NA 24.3350000         NA 2.534378   2.52813
#> 4   12800  5252   632  4769   632  8.3101266  7.5458861 2.534378   2.52813
#> 5   13600  5648  1779 10562  3200  3.1748173  3.3006250 2.534378   2.52813
#> 6   21400  2516   891  2516   891  2.8237935  2.8237935 2.534378   2.52813
#> 7   21500  4877  1277  4877  1277  3.8191073  3.8191073 2.534378   2.52813
#> 8   21700 14450   937 14450   937 15.4215582 15.4215582 2.534378   2.52813
#> 9   21900 28015  2814 28000  2814  9.9555792  9.9502488 2.534378   2.52813
#> 10  22000 16708  2762 16708  1960  6.0492397  8.5244898 2.534378   2.52813
#> 11  23000 13287   246 15531  1417 54.0121951 10.9604799 2.534378   2.52813
#> 12  23600  8683   339  8683   339 25.6135693 25.6135693 2.534378   2.52813
#> 13  30100  9881 20904  9861 24555  0.4726847  0.4015883 2.534378   2.52813
#> 14  40200  7211  3634  6909  2820  1.9843148  2.4500000 2.534378   2.52813
#> 15  41700  7645  1855  9089   964  4.1212938  9.4284232 2.534378   2.52813
#> 16  42600  3780    40  3780    NA 94.5000000         NA 2.534378   2.52813
#> 17  42700 15973  3459 16127  3964  4.6178086  4.0683653 2.534378   2.52813
#> 18  50100  8724  1279  8724  1279  6.8209539  6.8209539 2.534378   2.52813
#> 19  52900  6021  2750  6020    NA  2.1894545         NA 2.534378   2.52813
#> 20  60500  9126  5336  9306  7056  1.7102699  1.3188776 2.534378   2.52813
#> 21  61900  3174   445  3174   445  7.1325843  7.1325843 2.534378   2.52813
#> 22  62400 19400   350 19400   350 55.4285714 55.4285714 2.534378   2.52813
#> 23  62600  6908  5061  6908  5061  1.3649476  1.3649476 2.534378   2.52813
#> 24  63300  2548   170  2548   170 14.9882353 14.9882353 2.534378   2.52813
#> 25  70200  4302  1392  3027     0  3.0905172         NA 2.534378   2.52813
#> 26  70600 10255   770  9946   710 13.3181818 14.0084507 2.534378   2.52813
#> 27  70900  2234  2807  1600  4096  0.7958675  0.3906250 2.534378   2.52813
#> 28  72000   150  4466   150  4466  0.0335871  0.0335871 2.534378   2.52813
#> 29  80500 16941   300 17376   300 56.4700000 57.9200000 2.534378   2.52813
#> 30  80600  8977 20495  8977 20530  0.4380093  0.4372625 2.534378   2.52813
#> 31  81500  1888  2040  1888  2040  0.9254902  0.9254902 2.534378   2.52813
#> 32  82100  2371  1411  3271  1411  1.6803685  2.3182140 2.534378   2.52813
#> 33  82600  5401   815  5120   815  6.6269939  6.2822086 2.534378   2.52813
#> 34  83000  1670   250  2250   100  6.6800000 22.5000000 2.534378   2.52813
#> 35  83400  2600   400  2600   400  6.5000000  6.5000000 2.534378   2.52813
#> 36  91900  2986   150  3746   100 19.9066667 37.4600000 2.534378   2.52813
#> 37  93800   540  2224   540  2224  0.2428058  0.2428058 2.534378   2.52813
#> 38 100100 21802 13728 21802 24620  1.5881410  0.8855402 2.534378   2.52813
#> 39 110100  2744  2670  1839  1968  1.0277154  0.9344512 2.534378   2.52813
#> 40 110200 33029  9991 24000    NA  3.3058753         NA 2.534378   2.52813
#> 41 110300 18953  5185 19132  6655  3.6553520  2.8748310 2.534378   2.52813
#> 42 110600  8499  8938  8499  8938  0.9508839  0.9508839 2.534378   2.52813
#> 43 112400 12100  1200 14371  1150 10.0833333 12.4965217 2.534378   2.52813
#> 44 112900  1647    18  1647    18 91.5000000 91.5000000 2.534378   2.52813
#> 45 113000  5947  2815  1116  1775  2.1126110  0.6287324 2.534378   2.52813
#> 46 116000  5023   450  5023   450 11.1622222 11.1622222 2.534378   2.52813
#> 47 120100 50336 24755 50336 21633  2.0333670  2.3268155 2.534378   2.52813
#> 48 122400  6134  2245  5834  1460  2.7322940  3.9958904 2.534378   2.52813
#> 49 122800  2480   929  5475   929  2.6695371  5.8934338 2.534378   2.52813
#> 50 123500  7161  1357  7161  1357  5.2770818  5.2770818 2.534378   2.52813
#> 51 124500  1748  1238  1748  1238  1.4119548  1.4119548 2.534378   2.52813
#> 52 124700  9399   244 11732  1768 38.5204918  6.6357466 2.534378   2.52813
#> 53 126300  1948  2654  1948  2654  0.7339864  0.7339864 2.534378   2.52813
#> 54 126600  1800   300  1800     0  6.0000000         NA 2.534378   2.52813
#> 55 140100  7279  3004  7279  3004  2.4231025  2.4231025 2.534378   2.52813
#> 56 141600  3384   474  3225    NA  7.1392405         NA 2.534378   2.52813
#> 57 143900  4057  2132  4057  2132  1.9029081  1.9029081 2.534378   2.52813
#> 58 150500  5040  5999  7660  2930  0.8401400  2.6143345 2.534378   2.52813
#> 59 151400  1607   367  1607   367  4.3787466  4.3787466 2.534378   2.52813
#> 60 152800  2146   210  2146   210 10.2190476 10.2190476 2.534378   2.52813
#> 61 153400  3302   506  4037   506  6.5256917  7.9782609 2.534378   2.52813
#> 62 157600  2130   350  2130     0  6.0857143         NA 2.534378   2.52813
#> 63 160100 14022 13517 14026 12571  1.0373604  1.1157426 2.534378   2.52813
#> 64 161200   965   656   965   656  1.4710366  1.4710366 2.534378   2.52813
#> 65 161700  4611   180  4611    NA 25.6166667         NA 2.534378   2.52813
#> 66 162000  3621  3000  3621     0  1.2070000         NA 2.534378   2.52813
#> 67 162700  2194   176  2194   176 12.4659091 12.4659091 2.534378   2.52813
#> 68 164000  1794   165  1794    NA 10.8727273         NA 2.534378   2.52813
#> 69 164400   660   350   810   629  1.8857143  1.2877583 2.534378   2.52813
#> 70 166300  3427   583  3741   583  5.8782161  6.4168096 2.534378   2.52813
#> 71 170300  6563  2652  6563  2652  2.4747360  2.4747360 2.534378   2.52813
#> 72 171400  7495  1926  7495  2473  3.8914849  3.0307319 2.534378   2.52813
#> 73 173600  1995   100  1895   208 19.9500000  9.1105769 2.534378   2.52813
#> 74 175000   910  2000   910  2000  0.4550000  0.4550000 2.534378   2.52813
#> 75 180400 15448  5273 15406  4965  2.9296416  3.1029204 2.534378   2.52813
#> 76 182400  8200   980  8200   980  8.3673469  8.3673469 2.534378   2.52813
#> 77 183300  4833   484  4713   903  9.9855372  5.2192691 2.534378   2.52813
#> 78 183600  1173   370  1173   370  3.1702703  3.1702703 2.534378   2.52813
#> 79 184000  1904  3121  1904  3121  0.6100609  0.6100609 2.534378   2.52813
#> 80 184100  2697  1826  2697  1826  1.4769989  1.4769989 2.534378   2.52813
#> 81 186800  3990   150  3990   315 26.6000000 12.6666667 2.534378   2.52813
#> 82 193800  1935  1695     0  1935  1.1415929  0.0000000 2.534378   2.52813
#> 83 202400   964    25   964    25 38.5600000 38.5600000 2.534378   2.52813
#>    ratioStr ratioStr2 lowerLimit upperLimit outlier strata ranking
#> 1  3.286724  2.827904   -1.75159   19.24886       0      1       6
#> 2  3.286724  2.827904   -1.75159   19.24886       0      1      51
#> 3  3.286724  2.827904   -1.75159   19.24886       1      1      73
#> 4  3.286724  2.827904   -1.75159   19.24886       0      1      59
#> 5  3.286724  2.827904   -1.75159   19.24886       0      1      38
#> 6  3.286724  2.827904   -1.75159   19.24886       0      1      34
#> 7  3.286724  2.827904   -1.75159   19.24886       0      1      41
#> 8  3.286724  2.827904   -1.75159   19.24886       0      1      70
#> 9  3.286724  2.827904   -1.75159   19.24886       0      1      61
#> 10 3.286724  2.827904   -1.75159   19.24886       0      1      49
#> 11 3.286724  2.827904   -1.75159   19.24886       1      1      79
#> 12 3.286724  2.827904   -1.75159   19.24886       1      1      74
#> 13 3.286724  2.827904   -1.75159   19.24886       0      1       5
#> 14 3.286724  2.827904   -1.75159   19.24886       0      1      26
#> 15 3.286724  2.827904   -1.75159   19.24886       0      1      43
#> 16 3.286724  2.827904   -1.75159   19.24886       1      1      83
#> 17 3.286724  2.827904   -1.75159   19.24886       0      1      45
#> 18 2.196443  2.261188   -1.75159   19.24886       0      2      56
#> 19 2.196443  2.261188   -1.75159   19.24886       0      2      29
#> 20 2.196443  2.261188   -1.75159   19.24886       0      2      23
#> 21 2.196443  2.261188   -1.75159   19.24886       0      2      57
#> 22 2.196443  2.261188   -1.75159   19.24886       1      2      80
#> 23 2.196443  2.261188   -1.75159   19.24886       0      2      17
#> 24 2.196443  2.261188   -1.75159   19.24886       0      2      69
#> 25 2.196443  2.261188   -1.75159   19.24886       0      2      36
#> 26 2.196443  2.261188   -1.75159   19.24886       0      2      68
#> 27 2.196443  2.261188   -1.75159   19.24886       0      2       9
#> 28 2.196443  2.261188   -1.75159   19.24886       0      2       1
#> 29 2.196443  2.261188   -1.75159   19.24886       1      2      81
#> 30 2.196443  2.261188   -1.75159   19.24886       0      2       3
#> 31 2.196443  2.261188   -1.75159   19.24886       0      2      11
#> 32 2.196443  2.261188   -1.75159   19.24886       0      2      22
#> 33 2.196443  2.261188   -1.75159   19.24886       0      2      54
#> 34 2.196443  2.261188   -1.75159   19.24886       0      2      55
#> 35 2.196443  2.261188   -1.75159   19.24886       0      2      52
#> 36 2.196443  2.261188   -1.75159   19.24886       1      2      71
#> 37 2.196443  2.261188   -1.75159   19.24886       0      2       2
#> 38 2.436586  2.422342   -1.75159   19.24886       0      3      21
#> 39 2.436586  2.422342   -1.75159   19.24886       0      3      13
#> 40 2.436586  2.422342   -1.75159   19.24886       0      3      39
#> 41 2.436586  2.422342   -1.75159   19.24886       0      3      40
#> 42 2.436586  2.422342   -1.75159   19.24886       0      3      12
#> 43 2.436586  2.422342   -1.75159   19.24886       0      3      63
#> 44 2.436586  2.422342   -1.75159   19.24886       1      3      82
#> 45 2.436586  2.422342   -1.75159   19.24886       0      3      28
#> 46 2.436586  2.422342   -1.75159   19.24886       0      3      66
#> 47 2.436586  2.422342   -1.75159   19.24886       0      3      27
#> 48 2.436586  2.422342   -1.75159   19.24886       0      3      33
#> 49 2.436586  2.422342   -1.75159   19.24886       0      3      32
#> 50 2.436586  2.422342   -1.75159   19.24886       0      3      46
#> 51 2.436586  2.422342   -1.75159   19.24886       0      3      18
#> 52 2.436586  2.422342   -1.75159   19.24886       1      3      77
#> 53 2.436586  2.422342   -1.75159   19.24886       0      3       8
#> 54 2.436586  2.422342   -1.75159   19.24886       0      3      48
#> 55 2.436586  2.422342   -1.75159   19.24886       0      3      30
#> 56 2.436586  2.422342   -1.75159   19.24886       0      3      58
#> 57 2.436586  2.422342   -1.75159   19.24886       0      3      25
#> 58 1.908605  2.550171   -1.75159   19.24886       0      4      10
#> 59 1.908605  2.550171   -1.75159   19.24886       0      4      44
#> 60 1.908605  2.550171   -1.75159   19.24886       0      4      64
#> 61 1.908605  2.550171   -1.75159   19.24886       0      4      53
#> 62 1.908605  2.550171   -1.75159   19.24886       0      4      50
#> 63 1.908605  2.550171   -1.75159   19.24886       0      4      14
#> 64 1.908605  2.550171   -1.75159   19.24886       0      4      19
#> 65 1.908605  2.550171   -1.75159   19.24886       1      4      75
#> 66 1.908605  2.550171   -1.75159   19.24886       0      4      16
#> 67 1.908605  2.550171   -1.75159   19.24886       0      4      67
#> 68 1.908605  2.550171   -1.75159   19.24886       0      4      65
#> 69 1.908605  2.550171   -1.75159   19.24886       0      4      24
#> 70 1.908605  2.550171   -1.75159   19.24886       0      4      47
#> 71 1.908605  2.550171   -1.75159   19.24886       0      4      31
#> 72 1.908605  2.550171   -1.75159   19.24886       0      4      42
#> 73 1.908605  2.550171   -1.75159   19.24886       1      4      72
#> 74 1.908605  2.550171   -1.75159   19.24886       0      4       4
#> 75 2.954898  2.704086   -1.75159   19.24886       0      5      35
#> 76 2.954898  2.704086   -1.75159   19.24886       0      5      60
#> 77 2.954898  2.704086   -1.75159   19.24886       0      5      62
#> 78 2.954898  2.704086   -1.75159   19.24886       0      5      37
#> 79 2.954898  2.704086   -1.75159   19.24886       0      5       7
#> 80 2.954898  2.704086   -1.75159   19.24886       0      5      20
#> 81 2.954898  2.704086   -1.75159   19.24886       1      5      76
#> 82 2.954898  2.704086   -1.75159   19.24886       0      5      15
#> 83 2.954898  2.704086   -1.75159   19.24886       1      5      78